Attempt GHC compatability
