+++ plugins/tactics/src/Ide/Plugin/Tactic/GHC.hs
@@ -1,11 +1,14 @@
-{-# LANGUAGE ViewPatterns #-}
+{-# LANGUAGE CPP             #-}
+{-# LANGUAGE PatternSynonyms #-}
+{-# LANGUAGE ViewPatterns    #-}
 
 module Ide.Plugin.Tactic.GHC where
 
 import Data.Maybe (isJust)
+import Development.IDE.GHC.Compat
+import OccName
 import TcType
 import TyCoRep
-import TyCon
 import Type
 import TysWiredIn (intTyCon, floatTyCon, doubleTyCon, charTyCon)
 import Unique
@@ -66,4 +69,38 @@ lambdaCaseable (splitFunTy_maybe -> Just (arg, res))
   | isJust (algebraicTyCon arg)
   = Just $ isJust $ algebraicTyCon res
 lambdaCaseable _ = Nothing
+
+fromPatCompat :: PatCompat pass -> Pat pass
+#if __GLASGOW_HASKELL__ <= 808
+type PatCompat = Pat
+fromPatCompat = id
+#elif
+type PatCompat = LPat
+fromPatCompat = unLoc
+#endif
+
+------------------------------------------------------------------------------
+-- | Should make sure it's a fun bind
+pattern TopLevelRHS :: OccName -> [PatCompat GhcTc] -> LHsExpr GhcTc -> Match GhcTc (LHsExpr GhcTc)
+pattern TopLevelRHS name ps body <-
+  Match _
+    (FunRhs (L _ (occName -> name)) _ _)
+    ps
+    (GRHSs _
+      [L _ (GRHS _ [] body)] _)
+
+getPatName :: PatCompat GhcTc -> Maybe OccName
+getPatName (fromPatCompat -> p0) =
+  case p0 of
+    VarPat  _ x   -> Just $ occName $ unLoc x
+    LazyPat _ p   -> getPatName p
+    AsPat   _ x _ -> Just $ occName $ unLoc x
+    ParPat  _ p   -> getPatName p
+    BangPat _ p   -> getPatName p
+    ViewPat _ _ p -> getPatName p
+#if __GLASGOW_HASKELL__ >= 808
+    SigPat  _ p _ -> getPatName p
+#endif
+    XPat    p     -> getPatName $ unLoc p
+    _             -> Nothing
